logging: Separate translation unit (because of namespace conflicts)
